requirements.txt
aiohttp==3.9.3
aiosignal==1.3.1
anyio==4.3.0
attrs==23.2.0
beautifulsoup4==4.12.3
boto3==1.34.80
botocore==1.34.80
cachetools==5.3.3
certifi==2024.2.2
chardet==3.0.4
charset-normalizer==3.3.2
decorator==4.4.2
frozenlist==1.4.1
google-api-core==2.18.0
google-api-python-client==2.125.0
google-auth==2.29.0
google-auth-httplib2==0.2.0
google-cloud-texttospeech==2.16.3
googleapis-common-protos==1.63.0
googletrans==3.1.0a0
grpcio==1.62.1
grpcio-status==1.62.1
h11==0.9.0
h2==3.2.0
hpack==3.0.0
hstspreload==2024.4.1
httpcore==0.9.1
httplib2==0.22.0
httpx==0.13.3
hyperframe==5.2.0
idna==2.10
imageio==2.34.0
imageio-ffmpeg==0.4.9
imgkit==1.2.3
jmespath==1.0.1
lxml==5.2.1
moviepy==1.0.3
multidict==6.0.5
numpy==1.26.4
outcome==1.3.0.post0
packaging==24.0
pillow==10.3.0
praw==7.7.1
prawcore==2.4.0
proglog==0.1.10
proto-plus==1.23.0
protobuf==4.25.3
pyasn1==0.6.0
pyasn1_modules==0.4.0
pyparsing==3.1.2
PySocks==1.7.1
pysrt==1.1.2
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
requests==2.31.0
rfc3986==1.5.0
rsa==4.9
s3transfer==0.10.1
selenium==4.19.0
six==1.16.0
sniffio==1.3.1
sortedcontainers==2.4.0
soupsieve==2.5
telegram==0.0.1
tqdm==4.66.2
trio==0.25.0
trio-websocket==0.11.1
typing_extensions==4.11.0
update-checker==0.18.0
uritemplate==4.1.1
urllib3==2.2.1
webdriver-manager==4.0.1
websocket-client==1.7.0
wsproto==1.2.0
yarl==1.9.4
